replace component dialog, close material-form when add material
