+++ apps/delivery/delivery/src/app/template/category-list/category-list.component.html
@@ -1,7 +1,4 @@
 <mat-nav-list>
-  <button mat-raised-button color="primary" (click)="addCategory()">
-    Add new Category
-  </button>
   <h2 mat-subheader>Categories :</h2>
   <div *ngFor="let template of (template | keyvalue)">
     <mat-list-item>
