+++ apps/delivery/delivery/src/app/template/category-list/category-list.component.ts
@@ -1,41 +1,8 @@
 import { ChangeDetectionStrategy, Component, Input, Output } from '@angular/core';
-import { Material, MaterialService, MaterialStore } from '../../material/+state';
+import { Material, MaterialStore } from '../../material/+state';
 import { TemplateService } from '../+state';
-import { MatDialog, MatDialogRef } from '@angular/material';
+import { MatDialog } from '@angular/material';
 import { EventEmitter } from '@angular/core';
-
-@Component({
-  selector: 'add-category-dialog',
-  template: `
-    <div mat-dialog-content>
-      <h2>Add a new category</h2>
-      <mat-form-field appearance="outline" fxLayoutAlign="center">
-        <mat-label>File name</mat-label>
-        <input #category matInput placeholder="New category name"/>
-      </mat-form-field>
-    </div>
-    <div mat-dialog-actions fxLayoutAlign="center">
-      <button mat-stroked-button color="primary" (click)="close()">Cancel</button>
-      <button mat-raised-button color="primary">Create</button>
-    </div>
-  `
-})
-export class AddCategoryDialogComponent {
-  constructor(
-    public dialogRef: MatDialogRef<AddCategoryDialogComponent>,
-    private service: MaterialService
-  ) {
-  }
-
-  // public addCategory(category: string) {
-  //   this.service.addMaterial(category);
-  //   this.close();
-  // }
-
-  public close(): void {
-    this.dialogRef.close();
-  }
-}
 
 @Component({
   selector: 'category-list',
@@ -63,15 +30,7 @@ export class CategoryListComponent {
     this.materialStore.setActive(materialsIds);
   }
 
-  public addCategory(): void {
-    this.dialog.open(AddCategoryDialogComponent, {
-      width: '400px'
-    });
-  }
-
   public saveTemplate() {
     this.service.saveTemplate();
   }
-
 }
-
