+++ apps/delivery/delivery/src/app/template/template-view/template-view.component.scss
